Split putit's source files
