+++ src/putit/archive.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "binary_archive.hpp"
