+++ src/putit/archive_base.hpp
@@ -1,0 +1,21 @@
+#pragma once
+
+
+
+namespace elona
+{
+namespace putit
+{
+
+class IArchiveBase
+{
+};
+
+
+
+class OArchiveBase
+{
+};
+
+} // namespace putit
+} // namespace elona
